requirements.txt
pefile
pyelftools
